Move useHostLogin to a common file
